+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `mypy` for targets package
 
 ### Changed
-- Renamed `bounds_transform_func` target attribute to `transform_mode`
+- Renamed `bounds_transform_func` target attribute to `transformation`
 - Moved and renamed target transform utility functions
 
 ### Removed